fix: updated menu
