+++ src/app/shared/ui/menu/index.ts
@@ -1,2 +1,3 @@
 export * from './menu.component';
 export * from './menu.service';
+export * from './menu.interface';
