+++ src/app/shared/ui/menu/menu.component.ts
@@ -3,21 +3,16 @@ import { MatListModule } from '@angular/material/list';
 import { MatExpansionModule } from '@angular/material/expansion';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatIconModule } from '@angular/material/icon';
-import { AsyncPipe, NgClass, NgFor, NgIf } from '@angular/common';
+import { NgClass, NgFor, NgIf } from '@angular/common';
 import { NavigationEnd, Router, RouterModule } from '@angular/router';
 import { filter, map } from 'rxjs';
 import { CookieService } from 'ngx-cookie-service';
 
+import { UserInitialsButtonComponent } from '../user-initials-button/user-initials-button.component';
 import { MenuService } from './menu.service';
 import { menuItems } from './menu-items';
-import { UserRole } from '@core/store/user.interfaces';
-import { UserInitialsButtonComponent } from '../user-initials-button/user-initials-button.component';
-import { UserMenuComponent, UserMenuInputs } from '../user-menu';
-
-export interface MenuInputs {
-  role: UserRole;
-  userData: Required<UserMenuInputs>;
-}
+import { UserMenuComponent } from '../user-menu';
+import { MenuInputs } from './menu.interface';
 
 @Component({
   selector: 'app-menu[user]',
@@ -35,7 +30,6 @@ export interface MenuInputs {
     RouterModule,
     UserInitialsButtonComponent,
     UserMenuComponent,
-    AsyncPipe,
   ],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
